Statement API setup

+++ app/src/main/java/com/example/fintrack/domain/model/Statement.kt
@@ -1,7 +1,7 @@
 package com.example.fintrack.domain.model
 
 data class Statement(
-    val expensePercentage: Int,
+    val expensePercentage: Double,
     val totalBalance: Int,
     val totalExpense: Int,
     val totalIncome: Int
